# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TokenTrimmer Gateway API",
    "version": "1.0.0",
    "description": "OpenAI-compatible + Anthropic-native + Gemini gateway that routes, caches, compresses, and meters every LLM request. Eligible events can be minted on demand as tamper-evident signed estimates (VCR/WFR); signature verification is not universal receipt, issuer, savings-math, or provider-invoice proof. Auth: a `tt_live_` key in the `Authorization: Bearer` header (never caller-supplied org).",
    "license": { "name": "Apache-2.0" }
  },
  "servers": [
    { "url": "https://api.tokentrimmer.com", "description": "Hosted gateway" },
    { "url": "http://localhost:8080", "description": "Self-hosted (tt-proxy)" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "description": "A tt_live_ (production) or tt_test_ (sandbox) API key." }
    }
  },
  "security": [ { "BearerAuth": [] } ],
  "paths": {
...
    "/v1/routes/{id}": {
      "get": { "summary": "Fetch a route", "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "200": { "description": "The route." } } },
      "delete": { "summary": "Delete the observed route generation", "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }, { "name": "expected_revision", "in": "query", "required": true, "schema": { "type": "integer", "minimum": 1 }, "description": "Positive revision from a recent route management read." } ], "responses": { "200": { "description": "Deleted." }, "400": { "description": "Missing or invalid revision." }, "404": { "description": "Route not found." }, "409": { "description": "Route generation changed; reload before retrying." } } }
    },
...
    "/v1/routes/{id}/savings": { "get": { "summary": "Windowed route savings with legacy and signed estimates", "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }, { "name": "hours", "in": "query", "schema": { "type": "integer", "default": 720 } } ], "responses": { "200": { "description": "The route savings, including a signed row-level estimate that preserves regressions." } } } },
    "/v1/spend": {
      "get": { "summary": "Spend-today + MTD + budget-remaining", "description": "Backs the MCP cost-control tools (get_spend_today, check_budget_remaining). 503 until a SpendSource is wired.", "responses": { "200": { "description": "The spend summary." }, "503": { "description": "Spend reporting not configured." } } }
    },
...
  }
}